+++ payment_lyragw/__manifest__.py
@@ -1,16 +1,16 @@
 # coding: utf-8
 #
 # Copyright © Lyra Network.
-# This file is part of Lyra plugin for Odoo. See COPYING.md for license details.
+# This file is part of Lyra Collect plugin for Odoo. See COPYING.md for license details.
 #
 # Author:    Lyra Network (https://www.lyra.com)
 # Copyright: Copyright © Lyra Network
 # License:   http://www.gnu.org/licenses/agpl.html GNU Affero General Public License (AGPL v3)
 
 {
-    'name': 'Lyra Payment Acquirer',
+    'name': 'Lyra Collect Payment Acquirer',
     'version': '1.1.0',
-    'summary': 'Accept payments with Lyra secure payment gateway.',
+    'summary': 'Accept payments with Lyra Collect secure payment gateway.',
     'category': 'Accounting',
     'author': 'Lyra Network',
     'website': 'https://www.lyra.com/',