+++ payment_lyragw/controllers/__init__.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 #
 # Copyright © Lyra Network.
-# This file is part of Lyra plugin for Odoo. See COPYING.md for license details.
+# This file is part of Lyra Collect plugin for Odoo. See COPYING.md for license details.
 #
 # Author:    Lyra Network (https://www.lyra.com)
 # Copyright: Copyright © Lyra Network