+++ payment_lyragw/models/card.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 #
 # Copyright © Lyra Network.
-# This file is part of Lyra plugin for Odoo. See COPYING.md for license details.
+# This file is part of Lyra Collect plugin for Odoo. See COPYING.md for license details.
 #
 # Author:    Lyra Network (https://www.lyra.com)
 # Copyright: Copyright © Lyra Network
@@ -12,7 +12,7 @@ from ..helpers import constants
 
 class LyragwCard(models.Model):
     _name = 'lyragw.card'
-    _description = 'Lyra payment card'
+    _description = 'Lyra Collect payment card'
     _rec_name = 'label'
     _order = 'label'