+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - Rename some module files to match Odoo standards.
 - Added shipping information in sent form.
 - Consider AUTHORISED_TO_VALIDATE transaction status as a pending status.
-- Added Lyra supported currencies.
+- Added Lyra Collect supported currencies.
 - Added license notices.
 
 0.9.1, 2017-05-01
@@ -23,4 +23,4 @@
 
 0.9.0, 2017-01-22
 =============
-- Initial Lyra payment module for Odoo 9 by Sudokeys.
+- Initial Lyra Collect payment module for Odoo 9 by Sudokeys.